refactor(react-gui): lift the scene tree's three interactions out of the pane
ScenePane was 876 lines, and three of the things in it were interactions
with invariants of their own rather than parts of the pane:

  useTreeDragDrop      the drop indicator is set only after
                       planSceneNodeMove accepts the (source, target,
                       orientation) combination, so what the user sees and
                       what the drop does cannot disagree -- and the
                       dragged node travels in a ref because dragover
                       cannot read dataTransfer
  useVisibilityButton  the eye says three things, not two: on, off, and
                       "on, but an ancestor is hiding this anyway", which
                       is why it needs the parent lookup
  useTreeKeyboardNav   list navigation plus what a tree adds -- Right opens
                       a closed row before stepping in, Left closes an open
                       one before stepping out -- plus a Delete branch
                       gated on `editingNodeId`, without which a Backspace
                       meant for the rename input deleted the node being
                       renamed

`treeContents` stays. It reads fourteen things and turns them into rows:
that is the pane arranging its content, and pulling it out would only
move a fourteen-parameter signature somewhere else.

Extracting the drag hook made one omission visible immediately: passing
`onMoveNode` was easy to forget once it was an argument rather than a
closure capture, and the existing drag-and-drop test caught it before
anything else did.
